Adding IsInline() Method
I have added IsInline() method to the Edge object. It is different from
IsEquivalent() method so that it only checks FromNode and ToNode
equivalence. It does not check differences in AM and LM scores. This
will help detect multiple edges, that could be combined later in the
lattice, specially for multiple AM results fusion.